test(agents): colocate prepared build lifetime coverage
Move the unchanged serialized-build case into the existing candidate-lifetime suite so the original runtime suite stays within its line limit. Preserve the same harness, assertions, and deadlines.

Refs #132156.
